Rename cbaesp.yml to envused.yml
This is the environment file that was literally used to generate the results for the paper, but has some extraneous packages that aren't necessary for the results in the paper. Uploaded a more minimal cbaesp.yml file which will allow for the same results without such a large package overhead.